floats

+++ video.h
@@ -18,7 +18,7 @@ typedef struct texture_s {
 
 
 int videorun(int argc, char * argv[]);
-double F_sin ( double angle );
+float F_sin (float angle);
 
 #endif
